refactor(api): move interface and funcs to api/v0/plugins
