+++ api/v0/plugins/registry.go
