+++ api/v0/plugins/registry_test.go
