+++ internal/plugins/loader.go
@@ -1,7 +1,7 @@
 package plugins
 
 import (
-	"github.com/indaco/semver-cli/api/plugins"
+	"github.com/indaco/semver-cli/api/v0/plugins"
 	"github.com/indaco/semver-cli/internal/config"
 )
